Revert "improve retry mechanism, avoiding rely on AWS SQS retry and instead o… (#1999)"
This reverts commit 9d5460f63482aa4b016459ce63016a716bd6cc1f.
