+++ terraform-aws-github-runner/main.tf
@@ -37,7 +37,7 @@ resource "aws_sqs_queue" "queued_builds" {
   fifo_queue                  = true
   content_based_deduplication = true
   max_message_size            = 2048
-  message_retention_seconds   = var.runners_scale_up_sqs_message_ret_s
+  message_retention_seconds   = var.runners_scale_up_sqs_max_retry * var.runners_scale_up_sqs_visibility_timeout + 100
   redrive_policy              = jsonencode({
     deadLetterTargetArn = aws_sqs_queue.queued_builds_dead_letter.arn
     maxReceiveCount     = var.runners_scale_up_sqs_max_retry
@@ -57,7 +57,7 @@ resource "aws_sqs_queue" "queued_builds_retry" {
   name                        = "${var.environment}-queued-builds-retry"
   visibility_timeout_seconds  = var.runners_scale_up_sqs_visibility_timeout
   max_message_size            = 2048
-  message_retention_seconds   = var.runners_scale_up_sqs_message_ret_s
+  message_retention_seconds   = var.runners_scale_up_sqs_max_retry * var.runners_scale_up_sqs_visibility_timeout + 100
   redrive_policy              = jsonencode({
     deadLetterTargetArn = aws_sqs_queue.queued_builds_retry_dead_letter.arn
     maxReceiveCount     = var.runners_scale_up_sqs_max_retry
