+++ terraform-aws-github-runner/modules/runners/lambdas/runners/src/scale-runners/scale-up.ts
@@ -1,4 +1,4 @@
-import { Metrics, ScaleUpMetrics } from './metrics';
+import { Metrics, ScaleUpMetrics, sendMetricsAtTimeout, sendMetricsTimeoutVars } from './metrics';
 import { Repo, getRepoKey } from './utils';
 import { RunnerType, RunnerInputParameters, createRunner } from './runners';
 import {
@@ -30,14 +30,17 @@ export class RetryableScalingError extends Error {
   }
 }
 
-export async function scaleUp(
-  eventSource: string,
-  payload: ActionRequestMessage,
-  metrics: ScaleUpMetrics,
-): Promise<void> {
-  if (eventSource !== 'aws:sqs') {
-    throw Error('Cannot handle non-SQS events!');
-  }
+export async function scaleUp(eventSource: string, payload: ActionRequestMessage): Promise<void> {
+  if (eventSource !== 'aws:sqs') throw Error('Cannot handle non-SQS events!');
+
+  const metrics = new ScaleUpMetrics();
+  const sndMetricsTimout: sendMetricsTimeoutVars = {
+    metrics: metrics,
+  };
+  sndMetricsTimout.setTimeout = setTimeout(
+    sendMetricsAtTimeout(sndMetricsTimout),
+    (Config.Instance.lambdaTimeout - 10) * 1000,
+  );
 
   const repo: Repo = {
     owner: payload.repositoryOwner,
@@ -46,69 +49,81 @@ export async function scaleUp(
 
   const errors = [];
 
-  if (await shouldSkipForRepo(repo, metrics)) {
-    metrics.skipRepo(repo);
-    return;
-  }
-
-  metrics.runRepo(repo);
-  metrics.run();
-
-  const runnerTypes = await getRunnerTypes(
-    {
-      owner: repo.owner,
-      repo: Config.Instance.enableOrganizationRunners ? Config.Instance.scaleConfigRepo : repo.repo,
-    },
-    metrics,
-  );
-  /* istanbul ignore next */
-  const runnerLabels = payload?.runnerLabels ?? Array.from(runnerTypes.keys());
-
-  // ideally we should only have one label specfied but loop so we can go through them all if there are multiple
-  // if no labels are found this should just be a no-op
-  for (const runnerLabel of runnerLabels) {
-    const runnerType = runnerTypes.get(runnerLabel);
-    if (runnerType === undefined) {
-      console.info(`Runner label '${runnerLabel}' was not found in config for ` + `${repo.owner}/${repo.repo}`);
-      continue;
+  try {
+    if (await shouldSkipForRepo(repo, metrics)) {
+      metrics.skipRepo(repo);
+      return;
     }
-    if (
-      await allRunnersBusy(runnerType.runnerTypeName, repo, runnerType.is_ephemeral, runnerType.max_available, metrics)
-    ) {
-      try {
-        const createRunnerParams: RunnerInputParameters = {
-          environment: Config.Instance.environment,
-          runnerConfig: (awsRegion: string) => {
-            return createRunnerConfigArgument(runnerType, repo, payload.installationId, metrics, awsRegion);
-          },
-          runnerType: runnerType,
-        };
-        if (Config.Instance.enableOrganizationRunners) {
-          createRunnerParams.orgName = repo.owner;
-        } else {
-          createRunnerParams.repoName = getRepoKey(repo);
-        }
-        const awsRegion = await createRunner(createRunnerParams, metrics);
-        if (Config.Instance.enableOrganizationRunners) {
-          metrics.runnersOrgCreate(repo.owner, runnerType.runnerTypeName, awsRegion);
-        } else {
-          metrics.runnersRepoCreate(repo, runnerType.runnerTypeName, awsRegion);
-        }
-      } catch (e) {
-        errors.push(e);
 
-        /* istanbul ignore next */
-        if (Config.Instance.enableOrganizationRunners) {
-          metrics.runnersOrgCreateFail(repo.owner, runnerType.runnerTypeName);
-        } else {
-          metrics.runnersRepoCreateFail(repo, runnerType.runnerTypeName);
-        }
-        /* istanbul ignore next */
-        console.error(`Error spinning up instance of type ${runnerType.runnerTypeName}: ${e}`);
+    metrics.runRepo(repo);
+
+    const runnerTypes = await getRunnerTypes(
+      {
+        owner: repo.owner,
+        repo: Config.Instance.enableOrganizationRunners ? Config.Instance.scaleConfigRepo : repo.repo,
+      },
+      metrics,
+    );
+    /* istanbul ignore next */
+    const runnerLabels = payload?.runnerLabels ?? Array.from(runnerTypes.keys());
+
+    // ideally we should only have one label specfied but loop so we can go through them all if there are multiple
+    // if no labels are found this should just be a no-op
+    for (const runnerLabel of runnerLabels) {
+      const runnerType = runnerTypes.get(runnerLabel);
+      if (runnerType === undefined) {
+        console.info(`Runner label '${runnerLabel}' was not found in config for ` + `${repo.owner}/${repo.repo}`);
+        continue;
       }
-    } else {
-      console.info('There are available runners, no new runners will be created');
+      if (
+        await allRunnersBusy(
+          runnerType.runnerTypeName,
+          repo,
+          runnerType.is_ephemeral,
+          runnerType.max_available,
+          metrics,
+        )
+      ) {
+        try {
+          const createRunnerParams: RunnerInputParameters = {
+            environment: Config.Instance.environment,
+            runnerConfig: (awsRegion: string) => {
+              return createRunnerConfigArgument(runnerType, repo, payload.installationId, metrics, awsRegion);
+            },
+            runnerType: runnerType,
+          };
+          if (Config.Instance.enableOrganizationRunners) {
+            createRunnerParams.orgName = repo.owner;
+          } else {
+            createRunnerParams.repoName = getRepoKey(repo);
+          }
+          const awsRegion = await createRunner(createRunnerParams, metrics);
+          if (Config.Instance.enableOrganizationRunners) {
+            metrics.runnersOrgCreate(repo.owner, runnerType.runnerTypeName, awsRegion);
+          } else {
+            metrics.runnersRepoCreate(repo, runnerType.runnerTypeName, awsRegion);
+          }
+        } catch (e) {
+          errors.push(e);
+
+          /* istanbul ignore next */
+          if (Config.Instance.enableOrganizationRunners) {
+            metrics.runnersOrgCreateFail(repo.owner, runnerType.runnerTypeName);
+          } else {
+            metrics.runnersRepoCreateFail(repo, runnerType.runnerTypeName);
+          }
+          /* istanbul ignore next */
+          console.error(`Error spinning up instance of type ${runnerType.runnerTypeName}: ${e}`);
+        }
+      } else {
+        console.info('There are available runners, no new runners will be created');
+      }
     }
+  } finally {
+    clearTimeout(sndMetricsTimout.setTimeout);
+    sndMetricsTimout.metrics = undefined;
+    sndMetricsTimout.setTimeout = undefined;
+    metrics.sendMetrics();
   }
 
   if (errors.length > 0) {
